+++ AtmelStudio/Drafts/StateMachineGarageDoor/StateMachineGarageDoor/main.c
@@ -21,7 +21,7 @@ Simple schematics:
                       |         PC4 | ----> Motor IN1
  GND -------./ .----> | PB3     PC5 | ----> Motor IN2
              _        |             |
- GND ------. | .----> | PB4     PC6 | ----> Relay --- Electro magnetic lock
+ GND ------. | .----> | PB4     PC6 | ----> Relay --- Electro magnetic lock (SOLENIDE)
                       |_____________|
  Legend:
      _
@@ -39,31 +39,31 @@ Timer calculator: https://www.ee-diary.com/2021/07/programming-atmega328p-in-ctc
 #include <util/delay.h>
 #include "settings.h"
 
-volatile uint8_t extraTime = 0;				//Init extraTime
-volatile char state = CLOSED;				//Initial state is Closed
-volatile uint8_t count = 0;					//Init of count variable for button deboucing
-volatile uint8_t buttonPressed = 0;			//Flag for pressed button
+volatile uint8_t extraTime;					//Init extraTime
+volatile uint8_t numOfAlarms;				//Number of alarms triggered
+volatile char state = LOCKED;				//Initial state is Closed
+volatile uint8_t count;						//Init of count variable for button deboucing
+volatile uint8_t buttonPressed;				//Flag for pressed button
 volatile uint8_t limit = 100;				//Used in button debounce ISR
 //volatile char RF;							//Flag, if we are coming to state machine from RF
 
 /* Set up 8-bit timer for debounce interrupt
-The formula to calculate value of count to be loaded in OCR0A register (or OCR0B register) is:
-C=(Fclk/2*N*Fw)?1 where:
-- Fclk is the microcontroller CPU clock frequency
-- N is the pre-scalar value and
-- Fw is the output wave frequency
-
-The above equation can be rewritten in the following form where we use the output square wave time period(Tw)
-instead of the frequency(Fw).
-
-C=(Fclk*Tw/2*N)?1
-For example, if we want to create square wave signal with time period of Tw=100 us, we get:
-C=(8000000*0.0001/2*8)-1 = 49 
-
-So for 500 us time period, we get:
-C=(8000000*0.0005/2*8)-1 = 249
-
-*/
+ * The formula to calculate value of count to be loaded in OCR0A register (or OCR0B register) is:
+ * C=(Fclk/2*N*Fw)?1 where:
+ * - Fclk is the microcontroller CPU clock frequency
+ * - N is the pre-scalar value and
+ * - Fw is the output wave frequency
+ * 
+ * The above equation can be rewritten in the following form where we use the output square wave time period(Tw)
+ * instead of the frequency(Fw).
+ * 
+ * C=(Fclk*Tw/2*N)?1
+ * For example, if we want to create square wave signal with time period of Tw=100 us, we get:
+ * C=(8000000*0.0001/2*8)-1 = 49 
+ * 
+ * So for 500 us time period, we get:
+ * C=(8000000*0.0005/2*8)-1 = 249
+ */
 void debounceTimerStart() {
 	OCR0A = 249;						//See formula above
 	TCCR0A |= (1 << WGM01); 			//Set CTC mode
@@ -83,24 +83,27 @@ void initTimer() {
 }
 
 /*
-Start the 16-bit timer.
-Each overflow for 1024 prescaler at 8MHz lasts 8.388736 seconds
-*/
+ * Start the 16-bit timer.
+ * Each overflow for 1024 pre-scaler at 8MHz lasts app 8 seconds.
+ * Each overflow for 256 pre-scaler at 8MHz lasts app 2 seconds.
+ */
 void startTimer() {
 	TCNT1 = 0;
-	TCCR1B |= (1 << CS12) | (1 << CS10);			//Start the timer. To stop it, just write 0 to these bits.
+	//TCCR1B |= (1 << CS12) | (1 << CS10);			//Start the timer. To stop it, just write 0 to these bits.
+	TCCR1B |= (1 << CS12);							//256 pre-scaler => 2 seconds per overflow
 }
 
 /*
-Stop the timer.
-*/
+ * Stop the timer.
+ */
 void stopTimer() {
-	TCCR1B &= ~((1 << CS12) | (1 << CS10));
+	//TCCR1B &= ~((1 << CS12) | (1 << CS10));
+	TCCR1B &= ~(1 << CS12);
 }
 
 /*
-Shortcut for stop+start timer.
-*/
+ * Shortcut for stop+start timer.
+ */
 void restartTimer() {
 	stopTimer();
 	startTimer();
@@ -135,13 +138,13 @@ Electro magnetic lock. Shall be released when the state switches to OPENING
 and held back when the state switches to CLOSED.
 */
 
-void lockRelease() {
+void unlock() {
 	OUTPUT_PORT |= (1 << LOCK_PIN);
 	_delay_ms(200);								//Wait a little bit before starting a motor
 }
 
-void lockHold() {
-	_delay_ms(200);								//Wait a little bit before holding the lock
+void lock() {
+	_delay_ms(200);								//Wait a little bit before locking
 	OUTPUT_PORT &= ~(1 << LOCK_PIN);
 }
 
@@ -170,15 +173,54 @@ int main(void) {
 	{
 		switch (state)
 		{
+			case LOCKED:
+				if ((!(INPUT_PIN & (1 << OPEN_BTN_PIN))) & buttonPressed) {		//If the Open button was pressed
+					startTimer();							//Timer for timeout
+					state = ONE;
+				}
+				break;
+			case ONE:
+				if ((!(INPUT_PIN & (1 << OPEN_BTN_PIN))) & buttonPressed) {		//If the Open button was pressed
+					restartTimer();
+					state = TWO;
+				}
+				break;
+			case TWO:
+				if ((!(INPUT_PIN & (1 << CLOSE_BTN_PIN))) & buttonPressed) {		//If the Close button was pressed
+					restartTimer();
+					state = THREE;
+				}
+				break;
+			case THREE:
+				if ((!(INPUT_PIN & (1 << CLOSE_BTN_PIN))) & buttonPressed) {		//If the Close button was pressed
+					restartTimer();
+					state = IDLE;
+				}
+				break;
+			case IDLE:
+				stopTimer();
+				//If the Open door switch is pressed
+				if ((!(INPUT_PIN & (1 << OPEN_SWITCH_PIN))) & buttonPressed) {
+					state = OPEN;
+				}
+				//If the Closed door switch is pressed
+				else if ((!(INPUT_PIN & (1 << CLOSE_SWITCH_PIN))) & buttonPressed) {
+					state = CLOSED;
+				}
+				//Anything else is an error - Alarm
+				else {
+					state = ALARM;
+				}
+				break;
 			case CLOSED:
 				OUTPUT_PORT |= (1 << CLOSE_LED_PIN);			//Turn on Closed LED, signaling the door is fully closed
 				OUTPUT_PORT &= ~(1 << OPEN_LED_PIN);			//Turn off Open LED
 				OUTPUT_PORT &= ~(1 << ALARM_LED_PIN);			//Turn off Alarm LED
-				lockHold();										//Hold the lock
+				lock();
 				
 				if ((!(INPUT_PIN & (1 << OPEN_BTN_PIN))) & buttonPressed) {		//If the Open button was pressed
 						OUTPUT_PORT &= ~(1 << CLOSE_LED_PIN);	//Turn off CLOSE_LED_PIN
-						lockRelease();							//Release the lock
+						unlock();							//Release the lock
 						motorOpen();							//Start the motor to open the door
 						startTimer();							//Timer for timeout
 						state = OPENING;
@@ -251,14 +293,14 @@ int main(void) {
 				
 				//If the Open button was pressed, change state to OPENING
 				if ((!(INPUT_PIN & (1 << OPEN_BTN_PIN))) & buttonPressed) {
-						motorOpen();				//Start the motor to open the door
+						motorOpen();
 						restartTimer();
 						state = OPENING;
 				}
 				
 				//If the Closed door switch was hit
 				if ((!(INPUT_PIN & (1 << CLOSE_SWITCH_PIN))) & buttonPressed) {
-						motorStop();				//Stop the motor
+						motorStop();
 						stopTimer();
 						state = CLOSED;
 				}
@@ -278,21 +320,21 @@ int main(void) {
 				OUTPUT_PORT |= (1 << ALARM_LED_PIN);			//Turn on just a red LED
 				OUTPUT_PORT &= ~(1 << OPEN_LED_PIN);
 				OUTPUT_PORT &= ~(1 << CLOSE_LED_PIN);
-				stopTimer();
+				startTimer();
 				
 				//If the Close button was pressed
 				if ((!(INPUT_PIN & (1 << CLOSE_BTN_PIN))) & buttonPressed) {
 						OUTPUT_PORT &= ~(1 << ALARM_LED_PIN);	//Turn off ALARM_LED_PIN
-						motorClose();							//Start motor to close the door
-						startTimer();
+						motorClose();
+						restartTimer();
 						state = CLOSING;
 				}
 				
 				//If the Open button was pressed
 				if ((!(INPUT_PIN & (1 << OPEN_BTN_PIN))) & buttonPressed) {	
 						OUTPUT_PORT &= ~(1 << ALARM_LED_PIN);	//Turn off ALARM_LED_PIN
-						motorOpen();							//Start motor to open the door
-						startTimer();
+						motorOpen();
+						restartTimer();
 						state = OPENING;
 				}
 				break;			
@@ -302,53 +344,81 @@ int main(void) {
 	} //end while
 } //end main
 
+/*
+ * Overflow vector for timeouts on 16-bit register.
+ * Changed from 1024 pre-scaler to 256 (8 seconds -> 2 second).
+ * Idea is to use this one also for LED blinking when in Opening or Closing state. (TBD)
+ * Added numOfAlarms variable - if the overflow gets triggered more then once, the state
+ * machine goes to LOCKED state.
+ */
 ISR(TIMER1_OVF_vect)
 {
 	extraTime++;
-	if (extraTime > 0) {				//0 = 8 seconds, 1 = 16 seconds etc.
-		extraTime = 0;					//reset extraTime
-		state = ALARM;
+	if (extraTime > 5) {
+		extraTime = 0;
+		numOfAlarms++;
+		if (numOfAlarms <= 2) {
+			state = ALARM;
+		} else {
+			state = LOCKED;
+		}
 	}
 }
-
+/*
+ * Compare vector for button de-bounce on 8-bit timer.
+ * With limit = 100, we get 50 ms.
+ * Implemented for:
+ * - Open and Close push buttons
+ * - Fully open and fully closed switches
+ * - Emergency motor stop
+ */
 ISR(TIMER0_COMPA_vect)
 {
 	if (!(INPUT_PIN & (1 << CLOSE_BTN_PIN))) {
 		count++;
-		if(count > limit) {				//Triggers every 500 us, so with 100 we get 50 ms 
+		if(count > limit) {
 			count = 0;
-			buttonPressed = 1;			//The button is pressed
+			buttonPressed = 1;
 		} else {
-			buttonPressed = 0;			//The button is not pressed (yet)
+			buttonPressed = 0;
 		}	
 	}
 	if (!(INPUT_PIN & (1 << OPEN_BTN_PIN))) {
 		count++;
-		if(count > limit) {				//Triggers every 500 us, so with 100 we get 50 ms
+		if(count > limit) {
 			count = 0;
-			buttonPressed = 1;			//The button is pressed
+			buttonPressed = 1;
 			} else {
-			buttonPressed = 0;			//The button is not pressed (yet)
+			buttonPressed = 0;
 		}
 	}
 	if (!(INPUT_PIN & (1 << OPEN_SWITCH_PIN))) {
 		count++;
-		if(count > limit) {				//Triggers every 500 us, so with 100 we get 50 ms
+		if(count > limit) {
 			count = 0;
-			buttonPressed = 1;			//The button is pressed
+			buttonPressed = 1;
 			} else {
-			buttonPressed = 0;			//The button is not pressed (yet)
+			buttonPressed = 0;
 		}
 	}
 	if (!(INPUT_PIN & (1 << CLOSE_SWITCH_PIN))) {
 		count++;
-		if(count > limit) {				//Triggers every 500 us, so with 100 we get 50 ms
+		if(count > limit) {
 			count = 0;
-			buttonPressed = 1;			//The button is pressed
+			buttonPressed = 1;
 			} else {
-			buttonPressed = 0;			//The button is not pressed (yet)
+			buttonPressed = 0;
 		}
-	}		
+	}
+	if (!(INPUT_PIN & (1 << MOTOR_STOP_PIN))) {
+		count++;
+		if(count > limit) {
+			count = 0;
+			buttonPressed = 1;
+			} else {
+			buttonPressed = 0;
+		}
+	}	
 }
 
 /*